Extract duplicate extraction logic to object-types-utils
