+++ packages/schema/parse/src/extract/imported-object-types.ts
@@ -1,12 +1,10 @@
+import { TypeInfo, createImportedObjectDefinition } from "../typeInfo";
 import {
-  PropertyDefinition,
-  TypeInfo,
-  ImportedObjectDefinition,
-  createImportedObjectDefinition,
-  createPropertyDefinition,
-  createScalarDefinition,
-  createArrayDefinition,
-} from "../typeInfo";
+  extractFieldDefinition,
+  extractListType,
+  extractNamedType,
+  State,
+} from "./object-types-utils";
 
 import {
   DocumentNode,
@@ -19,12 +17,6 @@ import {
   DirectiveNode,
   ValueNode,
 } from "graphql";
-
-interface State {
-  currentImport?: ImportedObjectDefinition;
-  currentProperty?: PropertyDefinition;
-  nonNullType?: boolean;
-}
 
 const visitorEnter = (typeInfo: TypeInfo, state: State) => ({
   ObjectTypeDefinition: (node: ObjectTypeDefinitionNode) => {
@@ -95,69 +87,25 @@ const visitorEnter = (typeInfo: TypeInfo, state: State) => ({
     );
 
     typeInfo.importedObjectTypes.push(importedType);
-    state.currentImport = importedType;
-  },
-  FieldDefinition: (node: FieldDefinitionNode) => {
-    const importDef = state.currentImport;
-
-    if (!importDef) {
-      return;
-    }
-
-    if (node.arguments && node.arguments.length > 0) {
-      throw Error(
-        `Imported types cannot have methods. See type "${importDef.name}"`
-      );
-    }
-
-    const property = createPropertyDefinition(node.name.value);
-
-    state.currentProperty = property;
-    importDef.properties.push(property);
+    state.currentType = importedType;
   },
   NonNullType: (_node: NonNullTypeNode) => {
     state.nonNullType = true;
   },
   NamedType: (node: NamedTypeNode) => {
-    const property = state.currentProperty;
-
-    if (!property) {
-      return;
-    }
-
-    const modifier = state.nonNullType ? "" : "?";
-
-    property.scalar = createScalarDefinition(
-      property.name,
-      modifier + node.name.value,
-      state.nonNullType
-    );
-    state.nonNullType = false;
+    extractNamedType(node, state);
   },
   ListType: (_node: ListTypeNode) => {
-    const property = state.currentProperty;
-
-    if (!property) {
-      return;
-    }
-
-    if (property.scalar) {
-      return;
-    }
-
-    property.array = createArrayDefinition(
-      property.name,
-      "TBD",
-      state.nonNullType
-    );
-    state.currentProperty = property.array;
-    state.nonNullType = false;
+    extractListType(state);
+  },
+  FieldDefinition: (node: FieldDefinitionNode) => {
+    extractFieldDefinition(node, state);
   },
 });
 
 const visitorLeave = (typeInfo: TypeInfo, state: State) => ({
   ObjectTypeDefinition: (_node: ObjectTypeDefinitionNode) => {
-    state.currentImport = undefined;
+    state.currentType = undefined;
   },
   FieldDefinition: (_node: FieldDefinitionNode) => {
     state.currentProperty = undefined;
